Rewrite the code from scratch
Now `compress()` uses *kind-of-a-trie*, this made the code a lot faster for inputs that are not much compressable and also for the ones that are!

`decompress()` didn't change much, if at all, as it was already fast.
A small bug that presented itself while decoding unicode characters split between two verbatim chunks was fixed.

Some breaking changes were introduced, tho:
- `compatto` is now a function that accepts an object as options, for now only the dictionary, that is used to create the trie, and returns an object containing `compress()` and `decompress()`;
- the default dictionary is not provided by default anymore, as it made the script heavier for people that don't want to use it;
- `create()` doesn't exist anymore.